fix: refactor agentTester and agentTest
