+++ messages/agentTest.md
@@ -1,0 +1,11 @@
+# invalidAgentTestConfig
+
+Must provide one of: [ name, mdPath, specPath ] when instantiating an AgentTest.
+
+# missingConnection
+
+Must provide an org connection to get agent test data based on an AiEvaluationDefinition API name.
+
+# missingTestSpecData
+
+The agent test is missing the required data to provide a test spec.
