+++ src/agentTester.ts
@@ -5,33 +5,11 @@
  * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
 
-import { mkdir, readFile, writeFile } from 'node:fs/promises';
-import { dirname, join } from 'node:path';
-import { Connection, Lifecycle, PollingClient, SfError, StatusResult } from '@salesforce/core';
+import { Connection, Lifecycle, PollingClient, StatusResult } from '@salesforce/core';
 import { Duration, env } from '@salesforce/kit';
-import { ComponentSetBuilder, DeployResult, RequestStatus } from '@salesforce/source-deploy-retrieve';
-import { parse, stringify } from 'yaml';
-import { XMLBuilder, XMLParser } from 'fast-xml-parser';
 import { MaybeMock } from './maybe-mock';
 import { decodeHtmlEntities } from './utils';
-import {
-  type AvailableDefinition,
-  type AgentTestStartResponse,
-  type AgentTestStatusResponse,
-  type AgentTestResultsResponse,
-  type AiEvaluationDefinition,
-  type TestSpec,
-} from './types.js';
-
-/**
- * Events emitted during agent test creation for consumers to listen to and keep track of progress.
- */
-export const AgentTestCreateLifecycleStages = {
-  CreatingLocalMetadata: 'Creating Local Metadata',
-  Waiting: 'Waiting for the org to respond',
-  DeployingMetadata: 'Deploying Metadata',
-  Done: 'Done',
-};
+import { type AgentTestStartResponse, type AgentTestStatusResponse, type AgentTestResultsResponse } from './types.js';
 
 /**
  * A service for testing agents using `AiEvaluationDefinition` metadata. Start asynchronous
@@ -58,15 +36,8 @@ export const AgentTestCreateLifecycleStages = {
 export class AgentTester {
   private maybeMock: MaybeMock;
 
-  public constructor(private connection: Connection) {
+  public constructor(connection: Connection) {
     this.maybeMock = new MaybeMock(connection);
-  }
-
-  /**
-   * List the AiEvaluationDefinitions available in the org.
-   */
-  public async list(): Promise<AvailableDefinition[]> {
-    return this.connection.metadata.list({ type: 'AiEvaluationDefinition' });
   }
 
   /**
@@ -182,133 +153,6 @@ export class AgentTester {
 
     return this.maybeMock.request<{ success: boolean }>('POST', url);
   }
-
-  /**
-   * Creates and deploys an AiEvaluationDefinition from a specification file.
-   *
-   * @param apiName - The API name of the AiEvaluationDefinition to create
-   * @param specFilePath - The path to the specification file to create the definition from
-   * @param options - Configuration options for creating the definition
-   * @param options.outputDir - The directory where the AiEvaluationDefinition file will be written
-   * @param options.preview - If true, writes the AiEvaluationDefinition file to <api-name>-preview-<timestamp>.xml in the current working directory and does not deploy it
-   *
-   * @returns Promise containing:
-   * - path: The filesystem path to the created AiEvaluationDefinition file
-   * - contents: The AiEvaluationDefinition contents as a string
-   * - deployResult: The deployment result (if not in preview mode)
-   *
-   * @throws {SfError} When deployment fails
-   */
-  public async create(
-    apiName: string,
-    specFilePath: string,
-    options: { outputDir: string; preview?: boolean }
-  ): Promise<{ path: string; contents: string; deployResult?: DeployResult }> {
-    const parsed = parse(await readFile(specFilePath, 'utf-8')) as TestSpec;
-    const lifecycle = Lifecycle.getInstance();
-    await lifecycle.emit(AgentTestCreateLifecycleStages.CreatingLocalMetadata, {});
-    const preview = options.preview ?? false;
-    // outputDir is overridden if preview is true
-    const outputDir = preview ? process.cwd() : options.outputDir;
-    const filename = preview
-      ? `${apiName}-preview-${new Date().toISOString()}.xml`
-      : `${apiName}.aiEvaluationDefinition-meta.xml`;
-    const definitionPath = join(outputDir, filename);
-
-    const builder = new XMLBuilder({
-      format: true,
-      attributeNamePrefix: '$',
-      indentBy: '    ',
-      ignoreAttributes: false,
-    });
-
-    const xml = builder.build({
-      AiEvaluationDefinition: {
-        $xmlns: 'http://soap.sforce.com/2006/04/metadata',
-        ...(parsed.description && { description: parsed.description }),
-        name: parsed.name,
-        subjectName: parsed.subjectName,
-        subjectType: parsed.subjectType,
-        ...(parsed.subjectVersion && { subjectVersion: parsed.subjectVersion }),
-        testCase: parsed.testCases.map((tc) => ({
-          expectation: [
-            {
-              expectedValue: tc.expectedTopic,
-              name: 'topic_sequence_match',
-            },
-            {
-              expectedValue: `[${(tc.expectedActions ?? []).map((v) => `"${v}"`).join(',')}]`,
-              name: 'action_sequence_match',
-            },
-            {
-              expectedValue: tc.expectedOutcome,
-              name: 'bot_response_rating',
-            },
-          ],
-          inputs: {
-            utterance: tc.utterance,
-          },
-          number: parsed.testCases.indexOf(tc) + 1,
-        })),
-      },
-    }) as string;
-    const finalXml = `<?xml version="1.0" encoding="UTF-8"?>\n${xml}`;
-    await mkdir(outputDir, { recursive: true });
-    await writeFile(definitionPath, finalXml);
-    if (preview)
-      return {
-        path: definitionPath,
-        contents: finalXml,
-      };
-
-    const cs = await ComponentSetBuilder.build({ sourcepath: [definitionPath] });
-    const deploy = await cs.deploy({ usernameOrConnection: this.connection });
-    deploy.onUpdate((status) => {
-      if (status.status === RequestStatus.Pending) {
-        void lifecycle.emit(AgentTestCreateLifecycleStages.Waiting, status);
-      } else {
-        void lifecycle.emit(AgentTestCreateLifecycleStages.DeployingMetadata, status);
-      }
-    });
-
-    deploy.onFinish((result) => {
-      // small deploys like this, 1 file, can happen without an 'update' event being fired
-      // onFinish, emit the update, and then the done event to create proper output
-      void lifecycle.emit(AgentTestCreateLifecycleStages.DeployingMetadata, result);
-      void lifecycle.emit(AgentTestCreateLifecycleStages.Done, result);
-    });
-
-    const result = await deploy.pollStatus({ timeout: Duration.minutes(10_000), frequency: Duration.seconds(1) });
-
-    if (!result.response.success) {
-      throw new SfError(result.response.errorMessage ?? `Unable to deploy ${result.response.id}`);
-    }
-
-    return { path: definitionPath, contents: finalXml, deployResult: result };
-  }
-}
-
-/**
- * Convert the raw, detailed test results to another format.
- *
- * @param results The detailed results from a test run.
- * @param format The desired format. One of: json, junit, or tap.
- * @returns
- */
-export async function convertTestResultsToFormat(
-  results: AgentTestResultsResponse,
-  format: 'json' | 'junit' | 'tap'
-): Promise<string> {
-  switch (format) {
-    case 'json':
-      return jsonFormat(results);
-    case 'junit':
-      return junitFormat(results);
-    case 'tap':
-      return tapFormat(results);
-    default:
-      throw new Error(`Unsupported format: ${format as string}`);
-  }
 }
 
 /**
@@ -345,179 +189,5 @@ export function normalizeResults(results: AgentTestResultsResponse): AgentTestRe
         expectedValue: decodeHtmlEntities(r.expectedValue),
       })),
     })),
-  };
-}
-
-async function jsonFormat(results: AgentTestResultsResponse): Promise<string> {
-  return Promise.resolve(JSON.stringify(results, null, 2));
-}
-
-async function junitFormat(results: AgentTestResultsResponse): Promise<string> {
-  const builder = new XMLBuilder({
-    format: true,
-    attributeNamePrefix: '$',
-    ignoreAttributes: false,
-  });
-
-  const testCount = results.testCases.length;
-  const failureCount = results.testCases.filter(
-    (tc) =>
-      ['error', 'completed'].includes(tc.status.toLowerCase()) && tc.testResults.some((r) => r.result === 'FAILURE')
-  ).length;
-  const time = results.testCases.reduce((acc, tc) => {
-    if (tc.endTime && tc.startTime) {
-      return acc + new Date(tc.endTime).getTime() - new Date(tc.startTime).getTime();
-    }
-    return acc;
-  }, 0);
-
-  const suites = builder.build({
-    testsuites: {
-      $name: results.subjectName,
-      $tests: testCount,
-      $failures: failureCount,
-      $time: time,
-      property: [
-        { $name: 'status', $value: results.status },
-        { $name: 'start-time', $value: results.startTime },
-        { $name: 'end-time', $value: results.endTime },
-      ],
-      testsuite: results.testCases.map((testCase) => {
-        const testCaseTime = testCase.endTime
-          ? new Date(testCase.endTime).getTime() - new Date(testCase.startTime).getTime()
-          : 0;
-
-        return {
-          $name: testCase.testNumber,
-          $time: testCaseTime,
-          $assertions: testCase.testResults.length,
-          failure: testCase.testResults
-            .map((r) => {
-              if (r.result === 'FAILURE') {
-                return { $message: r.errorMessage ?? 'Unknown error', $name: r.name };
-              }
-            })
-            .filter((f) => f),
-        };
-      }),
-    },
-  }) as string;
-
-  return Promise.resolve(`<?xml version="1.0" encoding="UTF-8"?>\n${suites}`.trim());
-}
-
-export function humanFriendlyName(name: string): string {
-  // topic_sequence_match, action_sequence_match, and bot_response_rating have all changed
-  // eventually we can remove them
-  switch (name) {
-    case 'topic_sequence_match':
-    case 'topic_assertion':
-      return 'Topic';
-    case 'action_sequence_match':
-    case 'actions_assertion':
-      return 'Action';
-    case 'bot_response_rating':
-    case 'output_validation':
-      return 'Outcome';
-    default:
-      return name;
-  }
-}
-
-async function tapFormat(results: AgentTestResultsResponse): Promise<string> {
-  const lines: string[] = [];
-  let expectationCount = 0;
-  for (const testCase of results.testCases) {
-    for (const result of testCase.testResults) {
-      const status = result.result === 'PASS' ? 'ok' : 'not ok';
-      expectationCount++;
-      lines.push(`${status} ${expectationCount} ${testCase.testNumber}.${result.name}`);
-      if (status === 'not ok') {
-        lines.push('  ---');
-        lines.push(`  message: ${result.errorMessage ?? 'Unknown error'}`);
-        lines.push(`  expectation: ${result.name}`);
-        lines.push(`  actual: ${result.actualValue}`);
-        lines.push(`  expected: ${result.expectedValue}`);
-        lines.push('  ...');
-      }
-    }
-  }
-
-  return Promise.resolve(`Tap Version 14\n1..${expectationCount}\n${lines.join('\n')}`);
-}
-
-function transformStringToArray(str: string | undefined): string[] {
-  try {
-    if (!str) return [];
-    // Remove any whitespace and ensure proper JSON format
-    const cleaned = str.replace(/\s+/g, '');
-    return JSON.parse(cleaned) as string[];
-  } catch {
-    return [];
-  }
-}
-
-function castArray<T>(value: T | T[]): T[] {
-  return Array.isArray(value) ? value : [value];
-}
-
-/**
- * Generate a test specification file in YAML format.
- * This function takes a test specification object, cleans it by removing undefined and empty string values,
- * converts it to YAML format, and writes it to the specified output file.
- *
- * @param spec - The test specification object to be converted to YAML.
- * @param outputFile - The file path where the YAML output should be written.
- * @throws {Error} - May throw an error if file operations fail.
- * @returns A Promise that resolves when the file has been written.
- */
-export async function writeTestSpec(spec: TestSpec, outputFile: string): Promise<void> {
-  // strip out undefined values and empty strings
-  const clean = Object.entries(spec).reduce<Partial<TestSpec>>((acc, [key, value]) => {
-    if (value !== undefined && value !== '') return { ...acc, [key]: value };
-    return acc;
-  }, {});
-
-  const yml = stringify(clean, undefined, {
-    minContentWidth: 0,
-    lineWidth: 0,
-  });
-  await mkdir(dirname(outputFile), { recursive: true });
-  await writeFile(outputFile, yml);
-}
-
-/**
- * Generates a TestSpec object from an AI Evaluation Definition XML file.
- *
- * @param path - The file path to the AI Evaluation Definition XML file.
- * @returns Promise that resolves to a TestSpec object containing the parsed evaluation definition data.
- * @description Reads and parses an XML file containing AIEvaluationDefinition, converting it into a structured TestSpec format.
- *
- * @throws {Error} If the file cannot be read or parsed.
- */
-export async function generateTestSpecFromAiEvalDefinition(path: string): Promise<TestSpec> {
-  const xml = await readFile(path, 'utf-8');
-  const parser = new XMLParser();
-  const parsed = parser.parse(xml) as AiEvaluationDefinition;
-  return {
-    name: parsed.AiEvaluationDefinition.name,
-    description: parsed.AiEvaluationDefinition.description,
-    subjectType: parsed.AiEvaluationDefinition.subjectType,
-    subjectName: parsed.AiEvaluationDefinition.subjectName,
-    subjectVersion: parsed.AiEvaluationDefinition.subjectVersion,
-    testCases: castArray(parsed.AiEvaluationDefinition.testCase).map((tc) => {
-      const expectations = castArray(tc.expectation);
-      return {
-        utterance: tc.inputs.utterance,
-        // TODO: remove old names once removed in 258 (topic_sequence_match, action_sequence_match, bot_response_rating)
-        expectedTopic: expectations.find((e) => e.name === 'topic_sequence_match' || e.name === 'topic_assertion')
-          ?.expectedValue,
-        expectedActions: transformStringToArray(
-          expectations.find((e) => e.name === 'action_sequence_match' || e.name === 'actions_assertion')?.expectedValue
-        ),
-        expectedOutcome: expectations.find((e) => e.name === 'bot_response_rating' || e.name === 'output_validation')
-          ?.expectedValue,
-      };
-    }),
   };
 }
