+++ test/agentTestResults.test.ts
@@ -1,0 +1,54 @@
+/*
+ * Copyright (c) 2024, salesforce.com, inc.
+ * All rights reserved.
+ * Licensed under the BSD 3-Clause license.
+ * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
+ */
+import { readFile } from 'node:fs/promises';
+import { expect } from 'chai';
+import { convertTestResultsToFormat } from '../src/agentTestResults';
+import type { AgentTestResultsResponse } from '../src/types';
+
+describe('junit formatter', () => {
+  it('should transform test results to JUnit format', async () => {
+    const raw = await readFile('./test/mocks/einstein_ai-evaluations_runs_4KBSM000000003F4AQ_results/4.json', 'utf8');
+    const input = JSON.parse(raw) as AgentTestResultsResponse;
+    const output = await convertTestResultsToFormat(input, 'junit');
+    expect(output).to.equal(`<?xml version="1.0" encoding="UTF-8"?>
+<testsuites name="Guest_Experience_Agent" tests="3" failures="1" time="30000">
+  <property name="status" value="COMPLETED"></property>
+  <property name="start-time" value="2025-01-07T12:00:00Z"></property>
+  <property name="end-time" value="2025-01-07T12:00:10.35Z"></property>
+  <testsuite name="1" time="10000" assertions="3"></testsuite>
+  <testsuite name="2" time="10000" assertions="3"></testsuite>
+  <testsuite name="3" time="10000" assertions="3">
+    <failure message="An Apex error occurred: System.CalloutException: Bad Response: System.HttpResponse[Status=Not Found, StatusCode=404]" name="bot_response_rating"></failure>
+  </testsuite>
+</testsuites>`);
+  });
+});
+
+describe('tap formatter', () => {
+  it('should transform test results to TAP format', async () => {
+    const raw = await readFile('./test/mocks/einstein_ai-evaluations_runs_4KBSM000000003F4AQ_results/4.json', 'utf8');
+    const input = JSON.parse(raw) as AgentTestResultsResponse;
+    const output = await convertTestResultsToFormat(input, 'tap');
+    expect(output).to.equal(`Tap Version 14
+1..9
+ok 1 1.topic_sequence_match
+ok 2 1.action_sequence_match
+ok 3 1.bot_response_rating
+ok 4 2.topic_sequence_match
+ok 5 2.action_sequence_match
+ok 6 2.bot_response_rating
+ok 7 3.topic_sequence_match
+ok 8 3.action_sequence_match
+not ok 9 3.bot_response_rating
+  ---
+  message: An Apex error occurred: System.CalloutException: Bad Response: System.HttpResponse[Status=Not Found, StatusCode=404]
+  expectation: bot_response_rating
+  actual: It looks like I am unable to check the weather. There's something wrong with the Weather Service. How else can I assist you?
+  expected: The answer should start by describing expected conditions, for example "clear skies" or "50% chance of rain" and conclude with a range of high and low temperatures in degrees fahrenheit.
+  ...`);
+  });
+});
